Composer moved to future-proof URL
The appropriate redirect has been set up.

The composer will now be on a subfolder dedicated to personal projects.
As it has become a project completely unrelated to the website, it is
now a static file and doesn't use any Hugo features. It will probably be
added to a more fitted repository, but I will keep it on the website for
now.