Delete privacy resources
It was slightly outdated and I haven't checked some of the websites in a
long time. I neither want to keep it up to date or have it outdated, so
even though I strongly dislike breaking a link, I am taking it down. One
may always recover it from Git history.