Embed font into CSS file
This avoids the extra request for a file that is very small, removing
the flash of unstyled logo when loading the webpage. On top of that,
after Gzip compression, the new CSS file takes less space than both
files before (even with the base64 encoding!), so no bandwidth is lost
due to this (and the CSS file is cached anyway).